More GUI additions
Added a Role page, where you can choose to login as an admin, customer,
or create an account. Needs small work.

Added a drop down for role choosing when creating an account in the
Register Panel.

In the SlotsPanel.java (probably oging to change the name) updated it so
it looks more like a user dashboard, with a main content box in the
middle right, and butttons on the left (needs work). Little to no
functionality for now, will add that later (making more of a skeleton to
work with).