Changes made to the GUI, some Admin and Customer Functions.
Admins are able to add/remove slots and set pricing to specific slots

Customer are able to Login, register vehicles, view slots made by
Admins, reserve a slot, view their ticket and reservations, and pay for
their ticket (simulated)

Work in progess
-----------------------
Slot map should change colors when they are available or not, slots
should show the type when cliking on the slot to reserve it, filter
slots doesnt work yet. Clients should be able to see the color change in
slots when another customer reserves a spot. Admins should be able to
set slots with certain types view all current reservations in the
system, might let admins click on the slots themselves to allow for type
setting. Slot map should change colors when they are available or not.